% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAwage_shiny.R
\name{DAwage_shiny}
\alias{DAwage_shiny}
\title{The wage prediction shiny app}
\usage{
DAwage_shiny()
}
\value{
A shiny app, containing a density plot for estimated hourly wage.
}
\description{
This function returns a shiny app, which allows users to input the basic information of a job, and
get the estimated hourly wage for the specific job. The shiny app requires users to import all the needed features
from the job description, then exports the estimated hourly wage (in USD, including 25\%, 50\%, 75\% quantiles of the 100 predictions from bootstrap).
}
\details{
All the input boxes and sliders should have the same data type as what the predict_wage function requires.
}
