% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_wage.R
\name{predict_wage}
\alias{predict_wage}
\title{The wage prediction function}
\usage{
predict_wage(desc, remote, loc, type, degree, data, review)
}
\arguments{
\item{desc}{A string containing job-related description}

\item{remote}{A string taking only "on site" or "Remote" as input (Case sensitive)}

\item{loc}{A string taking abbreviations for states in the U.S. or "unknown" as inputs (ex: CO, NC, DC...(Case sensitive))}

\item{type}{A string taking only "Contract", "Full time", "mixed", "Part time", "Per diem"(day worker), "Temporary" or "unknown" as input (Case sensitive)}

\item{degree}{A string taking only "Bachelor", "Master", "NotSpecified" or "PhD" as input (Case sensitive)}

\item{data}{A string taking only "Yes" or "No" as input (Case sensitive), "Yes" means this job is related to data analyst type}

\item{review}{A numeric value (the number of reviews for the company where the job is)}
}
\value{
predict_wage returns a list containing median predicted hourly wage in USD, minimum predicted hourly wage in USD, maximum predicted
hourly wage in USD as well as 100 predicted hourly wages (vector) using bootstrap method.
}
\description{
Function implementing the hourly wage in USD prediction function.

The function takes a combination of job-related string and numeric value as inputs and
predicts the hour wage in USD using imported information for the job. And the function utilizes AdaBoost
model with tuned parameters to predict hour wages in USD.
}
\details{
The inputs are expected to be numeric or in the string types and contain only one value for each input.
And remote, loc, type, degree as well as data has to pick one value from available options provided in the
parameter description section as input. Any violations of these assumptions will result in an error.
}
\examples{

predict_wage("Requires sql", "Remote", "NC", "mixed", "Master", "Yes", 33)
predict_wage("Requires python", "on site", "CO", "Full time", "PhD", "No", 100)

}
